chore: Move fileprivate extensions to separate file
